refactor(ui): move WebRTC signaling behind an adapter interface
Extract the browser-offerer WebSocket signaling flow from the device route
into ui/src/webrtc/signaling/ so a different signaling transport can be
added later without touching the route.

- types.ts: SignalingOptions / SignalingController / SignalingHook. An
  adapter is a hook that owns the RTCPeerConnection lifecycle and reports
  the peer and its state through the RTC store.
- useWebSocketSignaling.ts: the existing code moved as-is: peer creation
  with ICE config, offer/answer and ICE trickle over the signaling
  WebSocket, the legacy HTTP session fallback for old devices, reconnect
  backoff, failure handling and the overlay loading text.
- index.ts: useSignaling() resolves an adapter by kind; one entry today.

The route keeps what it attaches to a fresh peer (transceivers, data
channels, media stream assembly) through onPeerConnection, and passes
the ICE servers it already loaded. ConnectionFailedOverlay and the
other-session outlet still receive setupPeerConnection unchanged.

No behaviour change intended. tsc, oxlint and a device build are clean.
